zephyr-env.sh: alternative shells compatibility
The patch makes envrironment setup script compatible with most commonly
used shells: bash and zsh by removing bash specific constructs. In case
of zsh, there is no realiable way of guessing whether we're sourced or
not, so making sure that the correct method is used is up to the
user. The checks are preserved for bash though.

The next assumption is that the existing users have already renamed
their ~/zephyr-env_install.bash file to ~/.zephyrrc so the file gets
sourced as before. For new installations, assume there is no .bash
suffix in user env file.

Signed-off-by: Maciej Borzecki <maciek.borzecki@gmail.com>