pinmux/stm32f10x: add driver for STM32F1 pinmux
Add driver for pin control subsystem in STM32F1 MCU series. The driver
allows for pin mode configuration without the ability of pin remapping.

The driver implements a pinmux driver API, with custom `func` and `pin`
encoding in API calls. The caller is expected to use STM32PIN() helper
macro for encoding port/pin numbers.

While the driver allows for configuration of IO pin mode, the explicit
configuration of alternate function is not supported. The usual practice
is to consult the reference manual or STM32Cube for information on pin
configuration required for particular mode. The driver assumes that user
performs such action.

Whenever an IO pin is being enabled, the driver will automatically
enable the clock for corresponding port. The driver does not implement
disabling of port's clock as this has potentially disruptive, as such
such operation should be done explicitly in the code.

The pin control module needs to be initialized before any other modules,
but after clock_control. For this reason, the driver is initialized at
PRIMARY level, with priority set to 2.

Change-Id: I8cb746d0f3cad72cd50b3355fe6d93a9f469be25
Signed-off-by: Maciej Borzecki <maciek.borzecki@gmail.com>