pinmux/stm32: formalize use of stm32_pin_func_t
In preparation for adding new STM32 SoCs, the helper type
stm32_pin_func_t will be used encoding pin mode and alternate function
in a convenient numerical format. The patch adds a number of helper
macros for encoding/decoding and updates the driver functions to
indicate where an encoded pin is expected to be passed/returned.

Change-Id: I51e674b927f025d125838d6695a442e274353f02
Signed-off-by: Maciej Borzecki <maciek.borzecki@gmail.com>